requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
dataclasses-json==0.6.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.301
langsmith==0.0.40
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.26.0
openai==0.28.0
packaging==23.1
pydantic==2.4.0
pydantic_core==2.10.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.21
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
yarl==1.9.2
